==> 03 Data load.R <==
# 03 data
library(here)


############### Load "00 Data Import.R" shit
load(here("data-edited","Holdings of Foreign Long-term Securities by U.S. Residents.Rda"))
load(here("data-edited","Holdings of U.S. Long-term Securities by Foreign Residents.Rda"))
load(here("data-edited","Short-term U.S. securities held by foreigners.Rda"))
load(here("data-edited","Gross Foreign Purchases and Sales of Long-term Securities by Country.Rda"))
load(here("data-edited","Major Foreign Holders of U.S. Treasuries.Rda"))
load(here("data-edited","Net Foreign purchases of U.S. Treasury Bonds and Notes.Rda"))
load(here("data-edited","Net Foreign purchases of U.S. Gov't corp & Federally-sponsored Agency Bonds.Rda"))
load(here("data-edited","Net Foreign purchases of U.S. Corporate & Other Bonds.Rda"))
load(here("data-edited","Net Foreign purchases of U.S. Stocks.Rda"))
load(here("data-edited","U.S.  acquisitions of foreign stocks through swaps.Rda"))
load(here("data-edited","TIC Data (Across-U.S. Border Financial Flows).Rda"))


############### Load "10 US Summary.R" shit
load(here("data-edited","Foreign securities held by U.S. residents - outflows.Rda"))
load(here("data-edited", "Foreign securities held by U.S. residents - outflows (SUMMARY).Rda"))
load(here("data-edited","U.S. securities held by foreign residents.Rda"))
load(here("data-edited","U.S. securities held by foreign residents (SUMMARY).Rda"))


################ Load "01 US Econ Data.R"
load(here("data-edited","Trade Weighted U.S. Dollar Index.Rda"))
load(here("data-edited","Effective Federal Funds Rate.Rda"))
load(here("data-edited","Daily Treasury Yield Curve Rates.Rda"))
load(here("data-edited","SPX.Rda"))
load(here("data-edited","DJI.Rda"))
load(here("data-edited","NDX.Rda"))
load(here("data-edited","VIX.Rda"))
load(here("data-edited","U.S. 5-Year, 5-Year Forward Inflation Expectation Rate.Rda"))
load(here("data-edited","U.S. Bureau of Labor Statistics, Historical CPI-U.Rda"))


################ Load "02 AU Econ Data.R"
load(here("data-edited","Exchange Rates – Daily - 2018 to Current – F11.1.Rda"))
load(here("data-edited","TB Yields 1992-Current (Indicative Mid Rates of Australian Government Securities – F16).Rda"))
load(here("data-edited","TIB Yields 1992-Current (Indicative Mid Rates of Australian Government Securities – F16).Rda"))
load(here("data-edited","AGB Yields 1995-Current (Capital Market Yields – Government Bonds – F2).Rda"))
load(here("data-edited","ABS CPI Data (Seasonally Adjusted - A2325847F).Rda"))
load(here("data-edited","RBA Cash rate target - A02.Rda"))
load(here("data-edited","RBA Inflation Expectations – G3.Rda"))
